:hammer: allow base engine to have constructor that has more shared code
