+++ moban/main.py
@@ -8,8 +8,8 @@
     :license: MIT License, see LICENSE for more details
 
 """
-
 import sys
+import logging
 import argparse
 
 import moban.reporter as reporter
@@ -19,6 +19,8 @@ import moban.exceptions as exceptions
 from moban import plugins
 from moban.utils import merge, open_yaml
 from moban.hashstore import HASH_STORE
+
+LOG = logging.getLogger(__name__)
 
 
 def main():
@@ -43,6 +45,7 @@ def main():
             exceptions.MobanfileGrammarException,
         ) as e:
             reporter.report_error_message(str(e))
+            LOG.exception(e)
             sys.exit(constants.ERROR)
     else:
         try:
