+++ tests/integration_tests/test_command_line_options.py
@@ -3,7 +3,6 @@ import sys
 from shutil import copyfile
 
 from mock import patch
-from moban.main import main
 from nose.tools import eq_, raises, assert_raises
 
 
@@ -13,13 +12,9 @@ class TestCustomOptions:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.jinja2.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
     @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_custom_options(self, fake_template_doer):
@@ -35,6 +30,8 @@ class TestCustomOptions:
             "a.jj2",
         ]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.jj2", "config.yaml", "moban.output"
@@ -44,6 +41,8 @@ class TestCustomOptions:
     def test_minimal_options(self, fake_template_doer):
         test_args = ["moban", "-c", self.config_file, "-t", "a.jj2"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.jj2", "config.yaml", "moban.output"
@@ -53,11 +52,12 @@ class TestCustomOptions:
     def test_missing_template(self):
         test_args = ["moban", "-c", self.config_file]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
         self.patcher1.stop()
-        self.patcher2.stop()
         os.unlink(self.config_file)
 
 
@@ -67,18 +67,16 @@ class TestOptions:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.jinja2.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
     @patch("moban.jinja2.engine.Engine.render_to_file")
     def test_default_options(self, fake_template_doer):
         test_args = ["moban", "-t", "a.jj2"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.jj2", "data.yml", "moban.output"
@@ -88,11 +86,12 @@ class TestOptions:
     def test_no_argments(self):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
         self.patcher1.stop()
-        self.patcher2.stop()
         os.unlink(self.config_file)
 
 
@@ -100,6 +99,8 @@ class TestOptions:
 def test_missing_configuration():
     test_args = ["moban", "-t", "a.jj2"]
     with patch.object(sys, "argv", test_args):
+        from moban.main import main
+
         main()
 
 
@@ -114,18 +115,16 @@ class TestNoOptions:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.jinja2.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
     @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -140,6 +139,8 @@ class TestNoOptions:
     def test_single_command_with_a_few_options(self, fake_template_doer):
         test_args = ["moban", "-t", "abc.jj2", "-o", "xyz.output"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -163,6 +164,8 @@ class TestNoOptions:
             "xyz.output",
         ]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -179,13 +182,14 @@ class TestNoOptions:
     def test_single_command_without_output_option(self, fake_template_doer):
         test_args = ["moban", "-t", "abc.jj2"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
         os.unlink(self.config_file)
         os.unlink(self.data_file)
         self.patcher1.stop()
-        self.patcher2.stop()
 
 
 class TestNoOptions2:
@@ -199,18 +203,16 @@ class TestNoOptions2:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.jinja2.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
     @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -223,7 +225,6 @@ class TestNoOptions2:
 
     def tearDown(self):
         self.patcher1.stop()
-        self.patcher2.stop()
         os.unlink(self.config_file)
         os.unlink(self.data_file)
 
@@ -238,18 +239,16 @@ class TestCustomMobanFile:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.jinja2.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
     @patch("moban.jinja2.engine.Engine.render_to_files")
     def test_single_command(self, fake_template_doer):
         test_args = ["moban", "-m", self.config_file]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             call_args = list(fake_template_doer.call_args[0][0])
             eq_(
@@ -262,17 +261,18 @@ class TestCustomMobanFile:
 
     def tearDown(self):
         self.patcher1.stop()
-        self.patcher2.stop()
         os.unlink(self.config_file)
         os.unlink(self.data_file)
 
 
-@patch("moban.jinja2.engine.verify_the_existence_of_directories")
+@patch("moban.plugins.verify_the_existence_of_directories")
 def test_duplicated_targets_in_moban_file(fake_verify):
     config_file = "duplicated.moban.yml"
     copyfile(os.path.join("tests", "fixtures", config_file), ".moban.yml")
     test_args = ["moban"]
     with patch.object(sys, "argv", test_args):
+        from moban.main import main
+
         assert_raises(SystemExit, main)
     os.unlink(".moban.yml")
 
@@ -288,6 +288,8 @@ class TestInvalidMobanFile:
             f.write("")
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     @raises(SystemExit)
@@ -297,6 +299,8 @@ class TestInvalidMobanFile:
             f.write("not: related")
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     @raises(SystemExit)
@@ -306,6 +310,8 @@ class TestInvalidMobanFile:
             f.write("configuration: test")
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
 
     def tearDown(self):
@@ -322,33 +328,30 @@ class TestComplexOptions:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.jinja2.engine.verify_the_existence_of_directories"
-        )
-        self.patcher2 = patch(
-            "moban.engine_factory.verify_the_existence_of_directories"
+            "moban.plugins.verify_the_existence_of_directories"
         )
         self.patcher1.start()
-        self.patcher2.start()
 
-    @patch("moban.jinja2.engine.Engine.render_to_files")
-    def test_single_command(self, fake_template_doer):
+    def test_single_command(self):
         test_args = ["moban"]
         with patch.object(sys, "argv", test_args):
-            main()
-            call_args = list(fake_template_doer.call_args[0][0])
-            eq_(
-                call_args,
-                [
-                    ("README.rst.jj2", "custom-data.yaml", "README.rst"),
-                    ("setup.py.jj2", "data.yml", "setup.py"),
-                ],
-            )
+            from moban.main import main
+
+            with patch("moban.jinja2.engine.Engine.render_to_files") as fake:
+                main()
+                call_args = list(fake.call_args[0][0])
+                eq_(
+                    call_args,
+                    [
+                        ("README.rst.jj2", "custom-data.yaml", "README.rst"),
+                        ("setup.py.jj2", "data.yml", "setup.py"),
+                    ],
+                )
 
     def tearDown(self):
         os.unlink(self.config_file)
         os.unlink(self.data_file)
         self.patcher1.stop()
-        self.patcher2.stop()
 
 
 class TestTemplateTypeOption:
@@ -361,6 +364,8 @@ class TestTemplateTypeOption:
     def test_mako_option(self, fake_template_doer):
         test_args = ["moban", "-t", "a.mako"]
         with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
             main()
             fake_template_doer.assert_called_with(
                 "a.mako", "data.yml", "moban.output"
